infra/cdk/requirements.txt
aws-cdk-lib>=2.150.0
constructs>=10.3.0

